requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Motor drift testing (Raspberry Pi only)
RPi.GPIO>=0.7.1
matplotlib>=3.8.0